Check built-in flows through the production dev-build rewrite
Replaces the test-only staging machinery with the code the shell already
ships. BuiltInFlows.extracted on a dev version does the whole job — repins
every flow to the running version, inserts `//> using repository ivy2Local`,
materialises them, enumerated by the generated index — so the check now
exercises the real "a dev build runs built-ins against the locally published
build" path instead of a fourth copy of the pin rewrite, and no longer
carries a hardcoded flow list.

The sbt version read-back goes too: build.sbt injects
`-Dorca.build.version` into the forked test JVMs, so
`sbt publishLocal "<module>/testOnly ..."` publishes and injects the same
version by construction. FlowAuthoringSmokeTest drops its own
publishLocal + version parsing for the same property.

On a release version `extracted` takes the no-rewrite path and the flows
would resolve from Maven Central — where a mid-release artifact isn't
published yet — so the suite skips itself there; the commit a tag points at
was already checked as a dev build.
